requirements.txt
anyio==4.3.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.0.1
confection==0.1.4
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
fastapi==0.101.1
filelock==3.13.3
fsspec==2024.3.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.22.2
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
lit==18.1.2
MarkupSafe==2.1.5
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nicknames==0.1.6
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==1.24.1
packaging==24.0
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.3.0
preshed==3.0.9
pydantic==1.10.15
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.13.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.27.0
sympy==1.12
thinc==8.1.12
threadpoolctl==3.4.0
tokenizers==0.13.3
torch==2.0.0
torchvision==0.15.1
tqdm==4.66.2
transformers==4.28.1
triton==2.0.0
typer==0.7.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
wasabi==1.1.2
Wikipedia-API==0.6.0